+++ NEWS.md
@@ -1,3 +1,11 @@
+## revealjmg 2.1.2
+
+* Replace the original `rmarkdown::render_supporting_files()` to a new
+  `render_supporting_files_2()`, which copies missing and modified files
+  from the source directory to the destination. This helps keep the 
+  destination tree updated when I rebuild things like `reveal.js` themes
+  and re-render documents.
+
 ## revealjmg 2.1.1
 
 * Updated to handle putting date into the title slide and metadata.